requirements.txt
backports-abc=>0.5
certifi=>2017.11.5
chardet=>3.0.4
Django=>1.9
django-filter=>1.1.0
djangorestframework=>3.7.3
futures=>3.2.0
idna=>2.6
Jinja2=>2.10
Markdown=>2.6.10
MarkupSafe=>1.0
msgpack-python=>0.4.8
pycrypto=>2.6.1
PyYAML=>3.12
pyzmq=>16.0.3
requests=>2.18.4
salt
tornado
urllib3=>1.22
